requirements.txt
aiodns==3.0.0
aiohttp==3.8.1
aiomultiprocess==0.9.0
aiosignal==1.2.0
argcomplete==2.0.0
async-timeout==4.0.2
attrs==22.1.0
aws-cdk.assets==1.168.0
aws-cdk.aws-acmpca==1.168.0
aws-cdk.aws-apigateway==1.168.0
aws-cdk.aws-applicationautoscaling==1.168.0
aws-cdk.aws-autoscaling==1.168.0
aws-cdk.aws-autoscaling-common==1.168.0
aws-cdk.aws-autoscaling-hooktargets==1.168.0
aws-cdk.aws-certificatemanager==1.168.0
aws-cdk.aws-cloudformation==1.168.0
aws-cdk.aws-cloudfront==1.168.0
aws-cdk.aws-cloudwatch==1.168.0
aws-cdk.aws-codebuild==1.168.0
aws-cdk.aws-codecommit==1.168.0
aws-cdk.aws-codeguruprofiler==1.168.0
aws-cdk.aws-codepipeline==1.168.0
aws-cdk.aws-codestarnotifications==1.168.0
aws-cdk.aws-cognito==1.168.0
aws-cdk.aws-dynamodb==1.168.0
aws-cdk.aws-ec2==1.168.0
aws-cdk.aws-ecr==1.168.0
aws-cdk.aws-ecr-assets==1.168.0
aws-cdk.aws-ecs==1.168.0
aws-cdk.aws-efs==1.168.0
aws-cdk.aws-eks==1.168.0
aws-cdk.aws-elasticloadbalancing==1.168.0
aws-cdk.aws-elasticloadbalancingv2==1.168.0
aws-cdk.aws-events==1.168.0
aws-cdk.aws-events-targets==1.168.0
aws-cdk.aws-globalaccelerator==1.168.0
aws-cdk.aws-iam==1.168.0
aws-cdk.aws-imagebuilder==1.168.0
aws-cdk.aws-kinesis==1.168.0
aws-cdk.aws-kinesisfirehose==1.168.0
aws-cdk.aws-kms==1.168.0
aws-cdk.aws-lambda==1.168.0
aws-cdk.aws-lambda-event-sources==1.168.0
aws-cdk.aws-logs==1.168.0
aws-cdk.aws-route53==1.168.0
aws-cdk.aws-route53-targets==1.168.0
aws-cdk.aws-s3==1.168.0
aws-cdk.aws-s3-assets==1.168.0
aws-cdk.aws-s3-notifications==1.168.0
aws-cdk.aws-sam==1.168.0
aws-cdk.aws-secretsmanager==1.168.0
aws-cdk.aws-servicediscovery==1.168.0
aws-cdk.aws-signer==1.168.0
aws-cdk.aws-sns==1.168.0
aws-cdk.aws-sns-subscriptions==1.168.0
aws-cdk.aws-sqs==1.168.0
aws-cdk.aws-ssm==1.168.0
aws-cdk.aws-stepfunctions==1.168.0
aws-cdk.aws-stepfunctions-tasks==1.168.0
aws-cdk.cloud-assembly-schema==1.168.0
aws-cdk.core==1.168.0
aws-cdk.custom-resources==1.168.0
aws-cdk.cx-api==1.168.0
aws-cdk.lambda-layer-awscli==1.168.0
aws-cdk.lambda-layer-kubectl==1.168.0
aws-cdk.lambda-layer-node-proxy-agent==1.168.0
aws-cdk.region-info==1.168.0
bc-python-hcl2==0.3.45
beautifulsoup4==4.11.1
boto3==1.24.51
botocore==1.27.53
cached-property==1.5.2
cachetools==5.2.0
cattrs==22.1.0
cdk-expects-matcher==0.1.2
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.1.0
checkov==2.1.137
click==8.1.3
click-option-group==0.5.3
cloudsplaining==0.5.0
colorama==0.4.4
ConfigArgParse==1.5.3
constructs==3.4.69
contextlib2==21.6.0
cyclonedx-python-lib==2.7.1
decorator==5.1.1
deep-merge==0.0.4
detect-secrets==1.3.0
docker==6.0.0
dockerfile-parse==1.2.0
docutils==0.16
dpath==1.5.0
exceptiongroup==1.0.0rc8
expects==0.9.0
frozenlist==1.3.1
gitdb==4.0.9
GitPython==3.1.27
idna==3.3
importlib-metadata==4.12.0
iniconfig==1.1.1
Jinja2==3.1.2
jmespath==1.0.1
jsii==1.64.0
jsonpath-ng==1.5.3
jsonschema==3.2.0
junit-xml==1.9
lark==1.1.2
Markdown==3.4.1
MarkupSafe==2.1.1
multidict==6.0.2
networkx==2.8.5
packageurl-python==0.10.1
packaging==21.3
pluggy==1.0.0
ply==3.11
policy-sentry==0.12.4
policyuniverse==1.5.0.20220613
prettytable==3.3.0
publication==0.0.3
py==1.11.0
pyasn1==0.4.8
pycares==4.2.2
pycep-parser==0.3.8
pycparser==2.21
pyparsing==3.0.9
pyrsistent==0.18.1
pytest==7.1.2
python-dateutil==2.8.2
PyYAML==5.4.1
regex==2022.8.17
requests==2.28.1
rsa==4.7.2
s3transfer==0.6.0
schema==0.7.5
semantic-version==2.10.0
six==1.16.0
smmap==5.0.0
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
tabulate==0.8.10
termcolor==1.1.0
toml==0.10.2
tomli==2.0.1
tqdm==4.64.0
typeguard==2.13.3
typing_extensions==4.3.0
update-checker==0.18.0
urllib3==1.26.11
wcwidth==0.2.5
websocket-client==1.3.3
yarl==1.8.1
zipp==3.8.1
